Handle BBC URLs passed to “Tune In URL”
If the user tries to pass in a URL along the lines of
https://www.bbc.co.uk/sounds/play/m000q3pg or https://www.bbc.co.uk/programmes/m000q3pg, handle that by mapping it to a `sounds://` URI.

A user might do that via the “Tune In URL” field in the web interface, or by sharing the URL to the LMS Material app using Andoirdʼs sharing feature.

Method `Slim::Player::ProtocolHandlers->registerURLHandler` was new in version 8.0 of LMS.  This change will do nothing on earlier versions.